[#1] Split bluetooth code
