+++ Sources/Blues/Bluetooth/BluetoothUtils.swift
@@ -1,58 +1,6 @@
 import IOBluetooth
 
-enum BluesError: Error {
-    case deviceNotFound(identifier: String)
-    case connectionError(identifier: String)
-}
-
-enum DeviceStatus {
-    case connected
-    case disconnected
-}
-
-struct Device {
-    private var device: IOBluetoothDevice
-
-    var name: String {
-        return device.name
-    }
-
-    var address: String {
-        return device.addressString.uppercased()
-    }
-
-    var status: DeviceStatus {
-        return device.isConnected() ? DeviceStatus.connected : DeviceStatus.disconnected
-    }
-
-    init(fromIOBluetoothDevice device: IOBluetoothDevice) {
-        self.device = device
-    }
-
-    func connect() throws -> Bool {
-        if device.isConnected() {
-            return false
-        }
-        let result = device.openConnection()
-        if result != kIOReturnSuccess {
-            throw BluesError.connectionError(identifier: device.name)
-        }
-        return true
-    }
-
-    func disconnect() throws -> Bool {
-        if !device.isConnected() {
-            return false
-        }
-        let result = device.closeConnection()
-        if result != kIOReturnSuccess {
-            throw BluesError.connectionError(identifier: device.name)
-        }
-        return true
-    }
-}
-
-class Bluetooth {
+class BluetoothUtils {
     private static func getPairedIOBluetoothDevices() -> [IOBluetoothDevice] {
         if let devices = IOBluetoothDevice.pairedDevices() as? [IOBluetoothDevice] {
             return devices
