+++ Sources/Blues/Errors.swift
@@ -1,0 +1,4 @@
+enum BluesError: Error {
+    case deviceNotFound(identifier: String)
+    case connectionError(identifier: String)
+}
